+++ src/config/Router.tsx
@@ -1,7 +1,9 @@
-import { createBrowserRouter, Form, RouterProvider } from 'react-router-dom'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import Layout from '../components/Layout'
 import NotFound from '../components/NotFound'
-import FormProject from '../components/form'
+import FormProject from '../views/NewProject'
+import Login from '../views/Login'
+import Register from '../views/Register'
 
 const Router = () => {
   const router = createBrowserRouter([
@@ -9,12 +11,23 @@ const Router = () => {
       path: '/',
       element: <Layout />,
       errorElement: <NotFound />,
+      children: [
+        {
+          path: 'new',
+          element: <FormProject />,
+        },
+      ],
     },
     {
-      path: '/new',
-      element: <FormProject />,
+      path: '/login',
+      element: <Login />,
       errorElement: <NotFound />,
-    }
+    },
+    {
+      path: '/register',
+      element: <Register />,
+      errorElement: <NotFound />,
+    },
   ])
   return <RouterProvider router={router} />
 }
